# spring/api-docs.yaml
openapi: 3.0.1
info:
  title: Конспект по языку Java
  description: Конспект по языку Java на русском языке
  contact:
    name: Токовенко Виктор
    url: https://github.com/Lorgar-Avrelian?tab=repositories
    email: victor-14-244@mail.ru
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /register:
    post:
      tags:
        - 1 Авторизация
        - Безопасность
      summary: Регистрация
      description: Пройти регистрацию
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDTO"
        required: true
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
      security:
        - basicAuth: []
...
  /login:
    post:
      tags:
        - 1 Авторизация
        - Безопасность
      summary: Войти
      description: Пройти авторизацию
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BasicAuthDTO"
        required: true
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      security:
        - basicAuth: []
...
  /set-role:
    patch:
      tags:
        - 1 Авторизация
        - Безопасность
      summary: Права
      description: Смена прав для зарегистрированного пользователя
      operationId: setRole
      parameters:
        - name: id
          in: query
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
        - name: role
          in: query
          description: Новая роль пользователя
          required: true
          schema:
            type: string
            enum:
              - ROLE_USER
              - ROLE_ADMIN
              - ROLE_OWNER
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginDTO"
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - basicAuth: []
  /set-password:
    patch:
      tags:
        - 1 Авторизация
        - Безопасность
      summary: Пароль
      description: Смена пароля для зарегистрированного пользователя
      operationId: setPassword
      parameters:
        - name: id
          in: query
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
        - name: password
          in: query
          description: Новый пароль пользователя
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginDTO"
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - basicAuth: []
...
  /users:
    get:
      tags:
        - 1 Авторизация
        - Безопасность
      summary: Пользователи
      description: Список зарегистрированных пользователей
      operationId: getUsers
      parameters:
        - name: page
          in: query
          description: Номер страницы
          required: false
          schema:
            type: integer
            format: int32
        - name: size
          in: query
          description: Количество пользователей на странице
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoginDTO"
        "400":
          description: Bad Request
        "403":
          description: Forbidden
      security:
        - basicAuth: []
...
  /csrf:
    get:
      tags:
        - 1 Авторизация
        - Безопасность
      summary: CSRF-токен
      description: Получить CSRF-токен
      operationId: csrf
      parameters:
        - name: csrfToken
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/CsrfToken"
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CsrfToken"
      security:
        - basicAuth: []
...
  /delete:
    delete:
      tags:
        - 1 Авторизация
        - Безопасность
      summary: Удалить
      description: Удалить пользователя
      operationId: delete
      parameters:
        - name: id
          in: query
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginDTO"
        "400":
          description: Bad Request
        "403":
          description: Forbidden
      security:
        - basicAuth: []
components:
  schemas:
...
    RegisterDTO:
      title: Регистрационные данные
      required:
        - login
        - password
        - passwordConfirmation
      type: object
      properties:
        login:
          title: Логин
          maxLength: 30
          minLength: 3
          type: string
          description: Логин пользователя
          default: user
        password:
          title: Пароль
          minLength: 3
          type: string
          description: Пароль пользователя
          default: "123"
        passwordConfirmation:
          title: Подтверждение пароля
          minLength: 3
          type: string
          description: Подтверждение пароля пользователя
          default: "123"
      description: "Логин, пароль и подтверждение пароля"
...
    BasicAuthDTO:
      title: Basic-аутентификация
      required:
        - login
        - password
      type: object
      properties:
        login:
          title: Логин
          maxLength: 30
          minLength: 3
          type: string
          description: Логин пользователя
          default: user
        password:
          title: Пароль
          minLength: 3
          type: string
          description: Пароль пользователя
          default: "123"
      description: Логин и пароль
...
    LoginDTO:
      title: Логин
      required:
        - id
        - login
        - role
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID пользователя
          format: int64
          default: 1
        login:
          title: Логин
          maxLength: 30
          minLength: 3
          type: string
          description: Логин пользователя
          default: user
        role:
          title: Роль
          maxLength: 30
          minLength: 3
          type: string
          description: Роль пользователя в системе
          default: USER
          enum:
            - ROLE_USER
            - ROLE_ADMIN
            - ROLE_OWNER
      description: Сущность логина пользователя
...
    CsrfToken:
      type: object
      properties:
        parameterName:
          type: string
        token:
          type: string
        headerName:
          type: string
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic